+++ src/components/BaseCard.tsx
