+++ src/components/Title.tsx
@@ -1,22 +1,5 @@
 import React from "react";
 
-interface TitleProps {
-  title: string;
-  subtitle?: string;
+export default function Title({ title }: { title: string }) {
+  return <div className="text-lg font-bold tracking-wider">{title}</div>;
 }
-
-export const Title: React.FunctionComponent<TitleProps> = ({
-  title,
-  subtitle,
-  children,
-}) => {
-  if (children) return <div className="section-title">{children}</div>;
-  if (subtitle)
-    return (
-      <div className="section-title">
-        <span className="pre-title">{title} </span>
-        {subtitle}
-      </div>
-    );
-  return <div className="section-title">{title}</div>;
-};
