+++ src/components/steps/SignAndPublishStep.tsx
