+++ src/index.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { App } from "./App2";
+import { App } from "./App";
 import "bootstrap/dist/css/bootstrap.min.css";
-import "./app.css";
 import "./index.css";
-import "./layout.css";
 import "./dappnode_styles.css";
 import "./dappnode_colors.css";
